requirements.txt
aniso8601==10.0.1
async-timeout==5.0.1
attrs==25.3.0
Automat==25.4.16
blinker==1.9.0
click==8.1.8
constantly==23.10.4
Flask==3.1.0
flask-cors==5.0.1
Flask-HTTPAuth==4.8.0
Flask-RESTful==0.3.10
greenlet==3.2.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.9.0
pymodbus==3.1.2
pyserial==3.5
pytz==2025.2
PyYAML==6.0.2
redis==5.3.0
six==1.17.0
SQLAlchemy==2.0.40
tomli==2.2.1
Twisted==24.11.0
typing_extensions==4.13.2
waitress==3.0.2
Werkzeug==3.1.3
zope.interface==7.2
